+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Arbitrary.hs
@@ -37,6 +37,38 @@ instance Crypto c => Arbitrary (ConwayDCert c) where
 -- Cardano.Ledger.Conway.Governance ------------------------------------------------------
 ------------------------------------------------------------------------------------------
 
+instance
+  (Era era, Arbitrary (PParams era), Arbitrary (PParamsUpdate era)) =>
+  Arbitrary (ConwayGovernance era)
+  where
+  arbitrary =
+    ConwayGovernance
+      <$> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+
+instance
+  (Era era, Arbitrary (PParams era), Arbitrary (PParamsUpdate era)) =>
+  Arbitrary (RatifyState era)
+  where
+  arbitrary =
+    RatifyState
+      <$> arbitrary
+      <*> arbitrary
+
+instance
+  (Era era, Arbitrary (PParams era), Arbitrary (PParamsUpdate era)) =>
+  Arbitrary (EnactState era)
+  where
+  arbitrary =
+    EnactState
+      <$> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+
 deriving instance (Era era, Arbitrary (PParamsUpdate era)) => Arbitrary (ConwayTallyState era)
 
 instance (Era era, Arbitrary (PParamsUpdate era)) => Arbitrary (GovernanceActionState era) where
@@ -98,9 +130,8 @@ instance
       <*> arbitrary
       <*> arbitrary
       <*> arbitrary
-      -- FIXME: For now this is turned off, rountrip tests in consensus are failing
-      <*> pure mempty -- arbitrary
-      <*> pure mempty -- arbitrary
+      <*> arbitrary
+      <*> arbitrary
 
 ------------------------------------------------------------------------------------------
 -- Cardano.Ledger.Conway.Rules -----------------------------------------------------------
@@ -132,6 +163,13 @@ instance (Era era, Arbitrary (PParamsUpdate era)) => Arbitrary (GovernanceProced
     oneof
       [ GovernanceVotingProcedure <$> arbitrary
       , GovernanceProposalProcedure <$> arbitrary
+      ]
+
+instance Era era => Arbitrary (ConwayTallyPredFailure era) where
+  arbitrary =
+    oneof
+      [ VoterDoesNotHaveRole <$> arbitrary <*> arbitrary
+      , GovernanceActionDoesNotExist <$> arbitrary
       ]
 
 instance